Basic support for the deno runner
