--- src/subcommands/dev.ts
@@ -1,239 +1,0 @@
-import { Command } from "@commander-js/extra-typings";
-import { SimpleLogger, text, Validator } from "@lmstudio/lms-common";
-import { EsPluginRunnerWatcher } from "@lmstudio/lms-es-plugin-runner/runner-watcher";
-import { UtilBinary } from "@lmstudio/lms-es-plugin-runner/util-binary";
-import { pluginManifestSchema } from "@lmstudio/lms-shared-types";
-import {
-  type LMStudioClient,
-  type PluginManifest,
-  type RegisterDevelopmentPluginOpts,
-} from "@lmstudio/sdk";
-import { type ChildProcessWithoutNullStreams } from "child_process";
-import { readFile } from "fs/promises";
-import { join } from "path";
-import { cwd } from "process";
-import { addCreateClientOptions, createClient } from "../createClient.js";
-import { exists } from "../exists.js";
-import { findProjectFolderOrExit } from "../findProjectFolder.js";
-import { addLogLevelOptions, createLogger } from "../logLevel.js";
-
-type PluginProcessStatus = "stopped" | "starting" | "running" | "restarting";
-interface PluginProcessOpts {
-  noNotify?: boolean;
-}
-
-class PluginProcess {
-  public constructor(
-    private readonly client: LMStudioClient,
-    private readonly registerDevelopmentPluginOpts: RegisterDevelopmentPluginOpts,
-    private readonly cwd: string,
-    private readonly logger: SimpleLogger,
-    private readonly opts: PluginProcessOpts = {},
-  ) {}
-  private readonly node = new UtilBinary("node");
-  private readonly args = ["--enable-source-maps", join(".lmstudio", "dev.js")];
-  private readonly serverLogger = new SimpleLogger("plugin-server", this.logger);
-  private readonly stderrLogger = new SimpleLogger("stderr", this.logger);
-
-  private currentProcess: ChildProcessWithoutNullStreams | null = null;
-  private status: PluginProcessStatus = "stopped";
-  private unregister: (() => Promise<void>) | null = null;
-  private firstTime = true;
-
-  private async startProcess() {
-    this.status = "starting";
-    const { unregister, clientIdentifier, clientPasskey } =
-      await this.client.plugins.registerDevelopmentPlugin(this.registerDevelopmentPluginOpts);
-    if (this.firstTime) {
-      const manifest = this.registerDevelopmentPluginOpts.manifest;
-      const identifier = `${manifest.owner}/${manifest.name}`;
-      if (this.opts.noNotify !== true) {
-        await this.client.system.notify({
-          title: `Plugin "${identifier}" started`,
-          description: "This plugin is run by lms CLI development server.",
-        });
-      }
-      this.firstTime = false;
-    }
-    this.unregister = unregister;
-    this.currentProcess = this.node.spawn(this.args, {
-      env: {
-        FORCE_COLOR: "1",
-        ...process.env,
-        LMS_PLUGIN_CLIENT_IDENTIFIER: clientIdentifier,
-        LMS_PLUGIN_CLIENT_PASSKEY: clientPasskey,
-      },
-      cwd: this.cwd,
-    });
-    this.currentProcess.stdout.on("data", data => this.logger.info(data.toString("utf-8").trim()));
-    this.currentProcess.stderr.on("data", data =>
-      this.stderrLogger.error(data.toString("utf-8").trim()),
-    );
-    this.currentProcess.on("exit", async (code, signal) => {
-      await this.unregister?.();
-      this.unregister = null;
-      if (code !== null) {
-        this.serverLogger.warn(`Plugin process exited with code ${code}`);
-      } else {
-        if (signal === "SIGKILL") {
-          // OK to ignore because we killed it
-        } else {
-          this.serverLogger.warn(`Plugin process exited with signal ${signal}`);
-        }
-      }
-      if (this.status === "restarting") {
-        this.startProcess();
-      } else {
-        this.status = "stopped";
-      }
-    });
-    this.status = "running";
-  }
-  public run() {
-    switch (this.status) {
-      case "stopped": {
-        this.startProcess();
-        break;
-      }
-      case "starting": {
-        // Already starting. Do nothing.
-        break;
-      }
-      case "running": {
-        this.status = "restarting";
-        if (this.unregister === null) {
-          this.currentProcess?.kill("SIGKILL");
-          this.currentProcess = null;
-        } else {
-          this.unregister().then(() => {
-            this.unregister = null;
-            this.currentProcess?.kill("SIGKILL");
-            this.currentProcess = null;
-          });
-        }
-        break;
-      }
-      case "restarting": {
-        // Already restarting. Do nothing.
-        break;
-      }
-    }
-  }
-}
-
-export const dev = addLogLevelOptions(
-  addCreateClientOptions(
-    new Command()
-      .name("dev")
-      .description("Starts the development server for the plugin in the current folder.")
-      .option(
-        "-i, --install",
-        text`
-          When specified, instead of starting the development server, installs the plugin to
-          LM Studio.
-        `,
-      )
-      .option(
-        "-y, --yes",
-        text`
-          Suppress all confirmations and warnings. Useful for scripting.
-          When used with --install, it will overwrite the plugin without asking.
-        `,
-      )
-      .option(
-        "--no-notify",
-        text`
-          When specified, will not produce the "Plugin started" notification in LM Studio.
-        `,
-      ),
-  ),
-).action(async options => {
-  const { install = false, yes = false, notify } = options;
-  const logger = createLogger(options);
-  const client = await createClient(logger, options);
-  const projectPath = await findProjectFolderOrExit(logger, cwd());
-  const manifestPath = join(projectPath, "manifest.json");
-  const manifestParseResult = pluginManifestSchema.safeParse(
-    JSON.parse(await readFile(manifestPath, "utf-8")),
-  );
-  if (!manifestParseResult.success) {
-    logger.error("Failed to parse the manifest file.");
-    logger.error(Validator.prettyPrintZod("manifest", manifestParseResult.error));
-    process.exit(1);
-  }
-  const manifest = manifestParseResult.data;
-
-  if (install) {
-    process.exit(await handleInstall(projectPath, manifest, logger, client));
-  } else {
-    await handleDevServer(projectPath, manifest, logger, client, { noNotify: !notify });
-  }
-});
-
-async function handleInstall(
-  projectPath: string,
-  manifest: PluginManifest,
-  logger: SimpleLogger,
-  client: LMStudioClient,
-): Promise<number> {
-  logger.info(`Installing the plugin ${manifest.owner}/${manifest.name}...`);
-  await client.repository.installLocalPlugin({ path: projectPath });
-  logger.info(`Successfully installed ${manifest.owner}/${manifest.name}.`);
-  return 0;
-}
-
-async function ensureNpmDependencies(path: string, logger: SimpleLogger, client: LMStudioClient) {
-  const packageJson = join(path, "package.json");
-  if (!(await exists(packageJson))) {
-    logger.error("No package.json found in the plugin folder.");
-    process.exit(1);
-  }
-  if (!(await exists(join(path, "node_modules")))) {
-    logger.info("Installing npm dependencies...");
-    await client.repository.installPluginDependencies(path);
-  } else {
-    logger.debug("node_modules already exists. Skipping npm install.");
-  }
-}
-
-async function handleDevServer(
-  projectPath: string,
-  manifest: PluginManifest,
-  logger: SimpleLogger,
-  client: LMStudioClient,
-  { noNotify }: { noNotify: boolean },
-) {
-  if (manifest.type !== "plugin") {
-    logger.error("The version of lms you are using only supports plugins.");
-    process.exit(1);
-  }
-
-  if (manifest.runner !== "ecmascript" && manifest.runner !== "node") {
-    logger.error("The version of lms you are using only supports node plugins.");
-    process.exit(1);
-  }
-
-  if (
-    manifest.type === "plugin" &&
-    (manifest.runner === "ecmascript" || manifest.runner === "node")
-  ) {
-    await ensureNpmDependencies(projectPath, logger, client);
-  }
-
-  logger.info(`Starting the development server for ${manifest.owner}/${manifest.name}...`);
-
-  const esbuild = new UtilBinary("esbuild");
-  const watcher = new EsPluginRunnerWatcher(esbuild, cwd(), logger);
-
-  const pluginProcess = new PluginProcess(client, { manifest }, projectPath, logger, { noNotify });
-
-  watcher.updatedEvent.subscribe(() => {
-    pluginProcess.run();
-  });
-  await watcher.start();
-
-  client.system.whenDisconnected().then(() => {
-    logger.info("Disconnected from the server. Stopping the development server.");
-    process.exit(1);
-  });
-}
